moving transpile tests from compiler/test_compiler.py to transpiler/test_transpile.py
